update deploy scripts for TLC
